+++ src/eval-frame/components/panes/declared-variable.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 
-import { ValueRenderer } from "../../../components/reps/value-renderer";
+import ValueRenderer from "../../../components/reps/value-renderer";
 
 export class DeclaredVariable extends React.Component {
   static propTypes = {
@@ -9,12 +9,11 @@ export class DeclaredVariable extends React.Component {
     varName: PropTypes.string
   };
   render() {
-    const r = true;
     return (
       <div className="declared-variable">
         <div className="declared-variable-name">{this.props.varName} = </div>
         <div className="declared-variable-value">
-          <ValueRenderer render={r} valueToRender={this.props.value} />
+          <ValueRenderer valueToRender={this.props.value} />
         </div>
       </div>
     );